+++ example/proj/foo.py
@@ -1,9 +1,18 @@
 def answer():
+    """
+    >>> answer()
+    42
+    """
     return 42
 
 class School():
 
     def food(self):
+        """
+        >>> s = School()
+        >>> s.food()
+        'awful'
+        """
         return 'awful'
 
     def age(self):